Add sentry middleware & sentry hub logger
